+++ Tests/ResetStackTests.swift
@@ -80,7 +80,7 @@ final class ResetStackTests: TestCase {
     func test_ThatPersistentStore_WithChanges_DoesNotHaveObjects_AfterReset() {
         // GIVEN: a stack and persistent store with data
         let model = CoreDataModel(name: modelName, bundle: modelBundle)
-        let factory = CoreDataStackFactory(model: model)
+        let factory = CoreDataStackProvider(model: model)
         let stack = try! factory.createStack().get()
         let context = stack.mainContext