+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This release closes the [9.0.0 milestone](https://github.com/jessesquires/JSQCor
 
 - Removed `StackResult` type in favor of Swift's new generic `Result` type.
 
+- Renamed `CoreDataStackFactory` to `CoreDataStackProvider`
+
 ### New
 
 - Added `typealias SaveResult = Result<NSManagedObjectContext, Error>` (Replaces the former `SaveResult` type.)